fix: readd challenge
